requirements.txt
python-telegram-bot[job-queue]>=20.7
pymongo>=4.6
python-dotenv>=1.0
mcp>=1.0
